Cleanup some more code
